Refactored how the model modifications are being cached and executed.  This will allow new modifications to be added easily (hopefully)
