+++ jaxb-ri/xjc/src/test/java/xjcTests/modelModifications/ModelModHandler.java
@@ -1,0 +1,12 @@
+package xjcTests.modelModifications;
+
+import com.sun.codemodel.JCodeModel;
+
+public interface ModelModHandler {
+
+	void handle(JCodeModel model) throws ModelModificationException;
+
+	public static class ModelModificationException extends Exception {
+
+	}
+}
